// src/app/shared/services/model/model.ts
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface userBookingInprogress {
  isSelected: boolean;
  id: number;
  bookingId: string;
  carName: string;
  deliveryStatus: string;
  rentalType: string;
  pickupDeliveryLocation1: string;
  pickupDeliveryLocation2: string;
  dropoffLocation1: string;
  dropoffLocation2: string;
  bookedOn: string;
  total: string;
  status: string;
  img: string;
}